# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /notanapi/profile/{id}/purchases:
    get:
      tags:
      - purchase-pagination-controller
      operationId: showMore
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagePurchase'
...
  /api/purchases/{id}:
    get:
      tags:
      - purchase-rest-controller
      summary: User gets a page of its purchases
      operationId: showMore_1
      parameters:
      - name: id
        in: path
        description: id of user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden. You have to be logged as a user to do this
        "400":
          description: Invalid user id and page number supplied
        "200":
          description: Found the page of purchases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "404":
          description: Page not found
  /api/purchases/price:
    get:
      tags:
      - purchase-rest-controller
      summary: Get price of all purchases to show it on a graph
      operationId: findPurchasesPrice
      responses:
        "404":
          description: Price of purchases not found
        "200":
          description: Found all prices of the purchases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "403":
          description: Forbidden. You have to be an admin to do this
  /api/purchases/id:
    get:
      tags:
      - purchase-rest-controller
      summary: Get id of all purchases to show it on a graph
      operationId: findPurchasesId
      responses:
        "404":
          description: Ids of purchases not found
        "403":
          description: Forbidden. You have to be an admin to do this
        "200":
          description: Found all ids of the purchases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
  /api/purchases/:
    get:
      tags:
      - purchase-rest-controller
      summary: Admin gets a page of all purchases
      operationId: showAdminMore_1
      responses:
        "200":
          description: Found the page of purchases
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "403":
          description: Forbidden. You have to be an admin to do this
        "400":
          description: Invalid page number supplied
        "404":
          description: Page not found
...
components:
  schemas:
    Dish:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        dishPrice:
          type: number
          format: float
        category:
          type: string
        nbuy:
          type: integer
          format: int32
        image:
          type: boolean
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameAllergen:
          type: string
        allergen:
          type: boolean
    Purchase:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        surname:
          type: string
        address:
          type: string
        postalCode:
          type: integer
          format: int32
        city:
          type: string
        country:
          type: string
        phoneNumber:
          type: integer
          format: int32
        price:
          type: number
          format: float
        dateDay:
          type: integer
          format: int32
        dateMonth:
          type: integer
          format: int32
        dateYear:
          type: integer
          format: int32
        current:
          type: boolean
        dishes:
          type: array
          items:
            $ref: '#/components/schemas/Dish'
...
    PagePurchase:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Purchase'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean